requirements.txt
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
msrestazure>=0.6.4
pyyaml>=6.0